refactor: Model API에 있던 학습된 모델 평가 서비스 Train API로 이동 및 잘못된 HTTP Methods 수정
