+++ app/router/model_router.py
@@ -2,7 +2,6 @@
 # -*- coding: utf-8 -*-
 # app/routes/model_router.py
 
-import os
 from io import StringIO
 
 import torch
@@ -15,7 +14,6 @@ from app.config.tibero import get_db
 from app.model.model import (
     Model,
     RequestInferenceImage,
-    RequestScore,
     get_inference_image_from_db,
     get_model_from_db,
 )
@@ -23,7 +21,6 @@ from app.model.train import RequestTrain, Train, get_train_by_id, new_train
 from app.util.source_generator import (
     get_dataloader_source,
     get_network_source,
-    get_test_metrics_source,
     get_train_source,
 )
 
@@ -100,46 +97,6 @@ def inference_image_model(
             return output
         else:
             return ",".join(map(str, output))
-
-
-@router.post("/{model_id}/test-metrics")
-async def test_metrics_model(
-    model_id: int,
-    req: RequestScore,
-    to_json: bool = False,
-    db: Connection = Depends(get_db),
-    kc: KernelClient = Depends(get_client),
-):
-    get_model_from_db(model_id, db)
-    train = get_train_by_id(req.train_id, db)
-
-    kernel = await kc.create_kernel()
-    if not kernel:
-        raise HTTPException(status_code=503, detail="no providers available")
-
-    model_filename = os.path.split(train.path)[1]
-    await kernel.send_file(train.path, model_filename)
-
-    result = await kernel.execute(
-        get_test_metrics_source(
-            model_filename,
-            req.testset.table_name,
-            req.testset.label_column_name,
-            req.testset.data_column_name,
-        ),
-        "Test model",
-    )
-
-    await kernel.stop()
-
-    try:
-        result = result[result.index("__RESULT__") + 1 :]
-        if to_json:
-            return result
-        else:
-            return "\n".join(result)
-    except:
-        raise HTTPException(status_code=400)
 
 
 def generate_source_response(source: str, filename: str) -> StreamingResponse:
